third_party: introduce new third party code layout, apply to STM32 HAL
As of today 3rd party code integration is a bit messy. We have some
submodules, some hard copies, some hard copies with patches and/or
in-tree build scripts...

The proposal in this patch is to follow this scheme:

- All 3rd party code should, in general, live in another repo
- Ideally, repos should be owned forks tracking their upstream
  counterparts, so if one they they disappear, change location, etc. we
  can rebuild the code
- Any necessary patches should happen in the forked repository, so it is
  clear on what we diverge. It is also easier to rebase against new
  releases.
- The folder structure should be:

  third_party/
  |- module/
  |  - wscript           # how we configure/build 3rd party code
  |  - any_glue_code.c   # any glue code (e.g. porting layer)
  |  - ore_glue_code.h   #
  \  - submodule         # git submodule

This patch applies this concept to the STM32 HAL. Note that a similar
concept is followed by Zephyr.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>