fw/freertos: introduce configASSERT_SAFE_TO_CALL_WAIT_FREERTOS_API
According to FreeRTOS docs, it is not safe to call non-ISR FreeRTOS
APIs if:

- You are in ISR (obviously)
- You are in a critical section

The former is because first, the API may allow blocking for a certain
amount of ticks (impossible within a critical section) and also because
it can yield to a another task which may not be in a critical section and
so interrupts would be enabled again. However, on Cortex-M in particular
yielding will not happen immediately until critical section is exited
(PendSV won't be processed until then). This means it should be safe to
call FreeRTOS APIs provided that the wait time is zero.

To relax the requirements on certain FreeRTOS APIs, a new helper is
introduced so that assertion will not be triggered if the call is
performed from a critical section and wait time is zero.

Ref. https://forums.freertos.org/t/
     critical-sections-freertos-api-calls/17352

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>