ci: use Docker image
Note that Docker image does not contain any Python dependencies because
as of today, some dependencies are held in the repository and installed
as editable. This makes things a bit trickier. We could potentially
cache pip cache if it becomes too slow, but still, we would not be able
to easily rely on cache-hit to not run pip install because of such
editable dependencies.

Note: see https://github.com/actions/runner-images/issues/6775 for the
git safe.directory requirement.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>