platform/asterix: configure display EXTCOMIN and deinit
PWM is not the most power-efficient way to toggle EXTCOMIN, using
RTC+PPI to toggle a pin yields far better results. However, we do not
care for now about this at the bootloader stage.

Also make sure both display SPI and PWM are left off before jumping to
the firmware.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>