fw/drivers: add stub files
Pebble driver layer is a bit of a shit show (it needs a huge refactor).
To boot a firmware it is assumed that quite a few APIs are implemented
and available, so bringing up a new platform can be challenging. Adding
stub files can help in this process, so that you can compile and link
successfully without unnecessary pain.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>