fw/drivers/flash/gd25lq255e: implement pseudo-protection
GD25LQ255E flash does not support block level locking operations, so
we cannot implement the protection APIs relying on SPI NOR capabilities.
The BP4-0 SR bits are not usable either because when disabling the write
protection, a chip erase (CE) is issued, making it a non-usable option
as we e.g. unlock PRF region to upgrade it.

BP4-0 bits could still be usable one day if we decide to finally lock
PRF forever and MFG_INFO regions located in the first MB of the flash.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>