freertos_application: reenable "cpustats" command [FIRM-264]
The 'cpustats' command provides incomplete data about system sleep, but
it does provide data.  In FIRM-264, we suspect that a divergence between
FreeRTOS and RTC ticks could be causing trouble for NimBLE, so we add a
mechanism to log FreeRTOS and RTC tick counts and see if they are lined
up with each other.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>