drivers/nrf5/hfxo: use clocksource.h, and have NimBLE disable stop mode
The real genesis of this commit is that NimBLE needs to disable stop
mode whenever it is using the HFXO to achieve predictable enough wake
latency.  This created a need for a separation between the NimBLE "turn
on the HFXO, and turn off stop mode" API, and the rest-of-PebbleOS "just
turn on the HFXO" API.  Of course, as it is said, the two hardest things
in computer science are cache invalidation, naming things, and
off-by-one errors -- and NimBLE occupies a fixed name for its API.

We rename the PebbleOS-exposed API to go with the clocksource.h APIs
that already exist for STM32, and provide the NimBLE-named API with the
capability to disable stop mode.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>