fw/drivers/task_watchdog: feed task watchdog from RTC
Unfortunately we're run out of low power timers (aka RTC) on Asterix. We
have RTC0 for BLE, RTC1 as a system RTC and RTC2 as a clock source to
generate the display EXTCOMIN signal.

This patch feeds the task watchdog from the system RTC driver IRQ, so
not at higher priority than the maximum FreeRTOS priority. It's not
ideal as it does not follow the implementation specs, but allows us to
still keep task watchdog operating. Task watchdog smells pretty bad
anyway, so it's not the end of the world.

If any other IRQ at the same priority as the RTC one takes long to
execute or IRQs are disabled for too long (both >1/2s), task watchdog feed
counter will be innacurate and its logic may not trigger in time to
provide some useful reports before the watchdog kicks us out. Unless I
miss something, this is not the most probable scenario.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>