fw/services/common/accel_manager: fix use-after-free
It looks like the Pebble accel manager came with a race that could occur
when start/stopping activity tracking (this happens e.g. when plugging the
charger).

The problem could occur if:

- prv_call_data_callback() is called for a state assigned to
  KernelBackground task (which defers to system task queue). At this
  point, event posted flag is set.
- Data from a previous event is consumed, so the event posted
  flag is cleared, ignoring if a new event was posted.
- Activity is stopped.
- Deferred free gets no pending events, so deferred free does not occur.
- Pending event comes and BOOM!

This patch uses a counter instead of a boolean for pending events, this
way, producers increase and consumers decrease and so avoid clobbering
a flag in case a new event is posted before the previous one is
processed. Accelerometer service assumes that the same task that
subscribes/unsubscribes is the one that does the processing, so no races
can occur in state access.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>