kernel/event_loop: avoid "overkill"ing watchfaces if KernelMain has gone out to lunch
The back-button-to-force-quit is handled on the KernelMain thread.  But if
shutting down the Launcher and entering a watchface takes a long time, and
the KernelMain thread gets starved, the back-button-to-force-quit sequence
could be activated even for just a short press of the back button!
Obviously it would be better if the system were not so darned slow, but in
the mean time, if it has to be, we can add a second check later to see if
the back-button-release event has fired by the time we actually get around
to killing the newborn watchface.

Fixes FIRM-425.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>