fix: apps need the --emit-relocs linker flag
The --emit-relocs linker flag is required to include a special ELF section
indicating where are the addresses that need relocation.

This elf section is converted at build time (inject_metadata.py) into a binary
table that is appended at the end of the app binary.

The process loader (`process_loader_storage.c`) reads this table to apply the
relocation offset to the app when it is loaded in memory.

With waf2, this linker flag (and a few less important ones) were ignored during
the build. This caused inject_metadata.py to inject 0 relocation entries and
caused some crashes. I saw this happened when a function pointer was passed to
applib but there were probably other bugs caused by this.

Update sdk/waftools/pebble_sdk_common.py

Signed-off-by: Thomas Sarlandie <thomas@sarlandie.net>