drivers/nrf5/qspi: poll instead of using a semaphore for very short reads
FreeRTOS semaphores and context switches can be pretty expensive, especially
in comparison to very short flash reads.  So for reads of low byte counts,
we do not actually try descheduling in the flash read mechanism -- instead,
we spin-wait on a flag that is set in the ISR.  It would be nice to fully
disable the ISR for these short reads, too, but that involves a little more
surgery moving to nrf instead of nrfx.  So for now, this is a big
improvement even still.

Performance improvements:

 * 4 bytes: 29us/read -> 20us/read (1.45x improvement)
 * 5 bytes: 50us/read -> 32us/read (1.56x improvement)
 * 16 bytes: 40us/read -> 24us/read (1.66x improvement)
 * 64 bytes: 51us/read -> 35us/read (1.45x improvement)
 * 256 bytes: 113us/read -> 83us/read (1.36x improvement)
 * 1024 bytes: 303us/read -> 303us/read (not affected by this change)

Signed-off-by: Joshua Wise <joshua@joshuawise.com>
Signed-off-by: Joshua Jun <joshuajun@proton.me>