drivers/nrf5/qspi: disable the IRQ when polling for very short reads
In 2480779da52a05be852543d8aff0204afbbb5187, we found that FreeRTOS
semaphores and context switches were very expensive.  They are, but it
is also very expensive just to have an IRQ at all (4us or so).  For
that, we avoid the nrfx_ API (which is inexorably committed to an IRQ,
once set up with a handler), and instead use the nrf_ API to turn off
the IRQ, set up the read, and spin-wait until it completes.

Performance improvements:

 * 4 bytes: 30us -> 25us (-5us)
 * 5 bytes: 42us -> 34us (-8us)
 * 16 bytes: 33us -> 29us (-4us)
 * 64 bytes: 45us -> 41us (-4us)
 * 256 bytes: 93us -> 89us (-4us)
 * 1024 bytes: 316us -> 315us (-1us; presumably noise)

Signed-off-by: Joshua Wise <joshua@accelerated.tech>