fw/drivers/sf32lb52/lptim: periodically calibrate
RC10K is wildly innacurate, and LPTIM driver assumed its frequency was
9000Hz. This patch measures RC10K (LPTIM clock) frequency at startup,
and adds a new call to configure a periodic timer to measure its
frequency, so we can tweak LPTIM compare register accordingly.

We should have a single point where to do the RC vs HXT48
measurements/calibrations as RTC driver also does that, but we can
refactor this a bit later.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>